+++ crypto/src/commonMain/kotlin/at/asitplus/awesn1/crypto/Pkcs8PrivateKeyInfo.kt
@@ -4,6 +4,7 @@
 package at.asitplus.awesn1.crypto
 
 import at.asitplus.awesn1.*
+import at.asitplus.awesn1.crypto.Pkcs1RsaPrivateKeyInfo.Companion.invoke
 import at.asitplus.awesn1.encoding.Asn1
 import at.asitplus.awesn1.serialization.Asn1Tag
 import at.asitplus.awesn1.serialization.DER
@@ -26,6 +27,7 @@ import kotlinx.serialization.Serializable
  * PrivateKey ::= OCTET STRING
  * Attributes ::= SET OF Attribute
  * ```
+ * @see Pkcs1RsaPrivateKeyInfo.of
  */
 @Serializable
 data class Pkcs8PrivateKeyInfo(
@@ -40,16 +42,15 @@ data class Pkcs8PrivateKeyInfo(
 
     override val pemLabel: String get() = PEM_LABEL_PRIVATE_KEY
 
-    @Throws(Asn1Exception::class)
-    fun decodeRsaPrivateKey(): Pkcs1RsaPrivateKeyInfo =
-        DER.decodeFromTlv(privateKey.asEncapsulatingOctetString().decodeRethrowing { next() })
+    @Deprecated("Moved to a more suitable location", ReplaceWith("Pkcs1RsaPrivateKeyInfo.of(this)"))
+    fun decodeRsaPrivateKey() =
+        Pkcs1RsaPrivateKeyInfo.of(this)
 
     @Throws(Asn1Exception::class)
     fun decodeEcPrivateKey(): Sec1EcPrivateKeyInfo =
         DER.decodeFromTlv(privateKey.asEncapsulatingOctetString().decodeRethrowing { next() })
 
     companion object : PemLabelSpec<Pkcs8PrivateKeyInfo> {
-        private val RSA_ENCRYPTION_OID = ObjectIdentifier("1.2.840.113549.1.1.1")
         private val EC_PUBLIC_KEY_OID = ObjectIdentifier("1.2.840.10045.2.1")
 
         const val PEM_LABEL_PRIVATE_KEY = "PRIVATE KEY"
@@ -60,15 +61,11 @@ data class Pkcs8PrivateKeyInfo(
         override val alternativePemLabels: Set<String> =
             setOf(PEM_LABEL_RSA_PRIVATE_KEY, PEM_LABEL_EC_PRIVATE_KEY)
 
+        @Deprecated("Moved to a more suitable location", ReplaceWith("Pkcs8PrivateKeyInfo(privateKey, attributes)"))
+        fun rsa(privateKey: Pkcs1RsaPrivateKeyInfo, attributes: Set<Asn1Element>? = null) =
+            this(privateKey, attributes)
 
-        fun rsa(privateKey: Pkcs1RsaPrivateKeyInfo, attributes: Set<Asn1Element>? = null): Pkcs8PrivateKeyInfo =
-            Pkcs8PrivateKeyInfo(
-                version = Version.V1,
-                privateKeyAlgorithm = X509AlgorithmIdentifier(RSA_ENCRYPTION_OID, Asn1.Null()),
-                privateKey = Asn1.OctetStringEncapsulating { +DER.encodeToTlv(privateKey) },
-                attributes = attributes,
-            )
-
+        @Deprecated("Moved to a more suitable location", ReplaceWith("Pkcs8PrivateKeyInfo(sec1Key, curveOid, attributes)"))
         fun ec(
             sec1Key: Sec1EcPrivateKeyInfo,
             curveOid: ObjectIdentifier?,